Incremental changes

+++ models/Position.cs
@@ -2,6 +2,6 @@ namespace AsteroidSharp.Models;
 
 public struct Position(int x, int y)
 {
-    public int X { get; set; } = x;
-    public int Y { get; set; } = y;
+    public float X { get; set; } = x;
+    public float Y { get; set; } = y;
 }
